Refactor:리팩토링 sns 객체 상수 파일로 분리
